1/26/2025 p1

+++ question_cpp/q.cpp
@@ -3,12 +3,12 @@ int main()
 {
     int n;
     int i;
-    long int a;
+    long long int a;
     long int r;
     scanf("%d%d%d",&a,&r,&n);
     for(i=1;n>i;i++)
     {
-        a = a*r;
+        a=a*r;
     }
     printf("%lld",a);
 }
